+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add DynamicExpresso library to use code evaluation for conditional attributes [[#74](https://github.com/DarkRewar/BaseTool/issues/74)]
 - Add `EnableIf` and `DisableIf` attributes to mark field as readonly is condition is valid or not [[#95](https://github.com/DarkRewar/BaseTool/issues/95)]
 - Add `Suffix` and `Prefix` attributes to add labels before or after a field in inspector [[#98](https://github.com/DarkRewar/BaseTool/issues/98)]
+- Add `ToSerializableDictionary` extension to `Dictionary` class [[#100](https://github.com/DarkRewar/BaseTool/issues/100)]
 
 ## 0.4.1